refactor test for `NonEmptyObject`  to new file
