+++ test/index.ts
@@ -15,7 +15,6 @@ import {
   DictionaryValues,
   Merge,
   MergeN,
-  NonEmptyObject,
   NonNever,
   noop,
   PickProperties,
@@ -831,26 +830,6 @@ function testNonNever() {
 
   type TestA = Assert<IsExact<keyof Mapped, "foo" | "bar" | "xyz">>;
   type TestB = Assert<IsExact<keyof NonNever<Mapped>, "foo" | "bar">>;
-}
-
-function testNonEmptyObject() {
-  type ObjectWithKeys = { foo: string; bar: number; xyz: undefined };
-  type EmptyObject = {};
-
-  type Cases = [
-    Assert<IsExact<NonEmptyObject<ObjectWithKeys>, ObjectWithKeys>>,
-    Assert<IsExact<NonEmptyObject<EmptyObject>, never>>,
-    // Works with Wrapper Object Types
-    Assert<IsExact<NonEmptyObject<String>, String>>,
-    Assert<IsExact<NonEmptyObject<Number>, Number>>,
-    Assert<IsExact<NonEmptyObject<Boolean>, Boolean>>,
-    // @ts-expect-error
-    Assert<IsExact<NonEmptyObject<string>, string>>, // Will Not work as T is a Primitive
-    // @ts-expect-error
-    Assert<IsExact<NonEmptyObject<number>, number>>, // Will Not work as T is a Primitive
-    // @ts-expect-error
-    Assert<IsExact<NonEmptyObject<undefined>, number>>, // Will Not work as T is a Primitive
-  ];
 }
 
 function testNonEmptyArray() {
